+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Headers:
 - Simplified module path to credential store implementations.
 - Montrose API updated to the 2026-06-11 version. Includes `CurrencyPosition`.
 - Renamed `Account` to `AccountHoldings`.
+- Renamed `HoldingsSelector` to `AccountFilter`.
 
 ## [0.0.2] - 2026-06-02